src/requirements.txt
pdfminer.six
unstructured[local-inference]
pi-heif
langchain
langchain-community
langchain-experimental
langchain_huggingface
sentence-transformers
transformers
python-docx
openpyxl
pandas
tiktoken
faiss-cpu
qdrant-client[fastembed]
langchain-qdrant
huggingface-hub
pytesseract
pdf2image
google-generativeai
scikit-learn
pdfplumber
rapidocr-onnxruntime
# Thư viện cho API
fastapi
uvicorn
python-multipart
python-dotenv